requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
psycopg2-binary
python-multipart
bcrypt<4.0.0
pytest
pytest-cov
httpx
